Refactor settings classes
