+++ src/dotnet/AIModel/ResourceProviders/AIModelResourceProviderService.cs
@@ -35,7 +35,7 @@ namespace FoundationaLLM.AIModel.ResourceProviders
     public class AIModelResourceProviderService(
         IOptions<InstanceSettings> instanceOptions,
         IAuthorizationService authorizationService,
-        [FromKeyedServices(DependencyInjectionKeys.FoundationaLLM_ResourceProvider_AIModel)] IStorageService storageService,
+        [FromKeyedServices(DependencyInjectionKeys.FoundationaLLM_ResourceProviders_AIModel)] IStorageService storageService,
         IEventService eventService,
         IResourceValidatorFactory resourceValidatorFactory,
         IServiceProvider serviceProvider,
