+++ src/dotnet/AuthorizationAPI/Program.cs
@@ -1,7 +1,5 @@
 using FoundationaLLM;
-using FoundationaLLM.Authorization.Interfaces;
 using FoundationaLLM.Authorization.Middleware;
-using FoundationaLLM.Authorization.Services;
 using FoundationaLLM.Common.Authentication;
 using FoundationaLLM.Common.Constants;
 using FoundationaLLM.Common.Constants.Configuration;
@@ -38,9 +36,9 @@ builder.AddCorsPolicies();
 
 // Add authentication configuration.
 builder.AddAuthenticationConfiguration(
-    KeyVaultSecretNames.FoundationaLLM_AuthorizationAPI_Entra_Instance,
-    KeyVaultSecretNames.FoundationaLLM_AuthorizationAPI_Entra_TenantId,
-    KeyVaultSecretNames.FoundationaLLM_AuthorizationAPI_Entra_ClientId,
+    AuthorizationKeyVaultSecretNames.FoundationaLLM_APIEndpoints_AuthorizationAPI_Configuration_Entra_Instance,
+    AuthorizationKeyVaultSecretNames.FoundationaLLM_APIEndpoints_AuthorizationAPI_Configuration_Entra_TenantId,
+    AuthorizationKeyVaultSecretNames.FoundationaLLM_APIEndpoints_AuthorizationAPI_Configuration_Entra_ClientId,
     null,
     policyName: "RequiredClaims",
     requireScopes: false,
@@ -48,7 +46,7 @@ builder.AddAuthenticationConfiguration(
 
 // Add OpenTelemetry.
 builder.AddOpenTelemetry(
-    KeyVaultSecretNames.FoundationaLLM_AuthorizationAPI_AppInsights_ConnectionString,
+    AuthorizationKeyVaultSecretNames.FoundationaLLM_APIEndpoints_AuthorizationAPI_AppInsightsConnectionString,
     ServiceNames.AuthorizationAPI);
 
 builder.Services.AddControllers();
