+++ src/dotnet/Common/Constants/Authorization/AuthorizationKeyVaultSecretNames.tt
@@ -6,38 +6,42 @@
 <#@ assembly name="System.Memory" #>
 <#@ import namespace="System.Text.Json" #>
 <#@ import namespace="System.Text.Json.Nodes" #>
-<#@ output extension=".json" #>
+<#@ output extension=".cs" #>
 <#
-    string inputFileName = this.Host.ResolvePath($"../Constants/Data/AppConfiguration.json");
+    string inputFileName = this.Host.ResolvePath($"../Data/AuthorizationAppConfiguration.json");
     string inputContent = File.ReadAllText(inputFileName);
     var jsonObjects = JsonNode.Parse(inputContent)!;
 
-    Func<JsonNode, string> getContentType = n => string.IsNullOrWhiteSpace(n["secret"].GetValue<string>())
-        ? n["content_type"].GetValue<string>()
-        : "application/vnd.microsoft.appconfig.keyvaultref+json;charset=utf-8";
+    Func<JsonNode, string> getName = n => n.GetValue<string>().Replace(":", "_");
 
     Func<JsonNode, string> getValue = n => string.IsNullOrWhiteSpace(n["secret"].GetValue<string>())
         ? n["value"].GetValue<string>()
         : $"{{\\\"uri\\\":\\\"${{env:AZURE_KEY_VAULT_ENDPOINT}}secrets/foundationallm-{n["secret"].GetValue<string>()}}}\\\"}}";
 #>
+namespace FoundationaLLM.Common.Constants.Configuration
 {
-    "items": [
-<#
+    /// <summary>
+    /// Defines all Azure Key vault secret names referred by the Azure App Configuration keys.
+    /// </summary>
+    public static partial class AuthorizationKeyVaultSecretNames
+    {<#
     foreach (var jsonObject in jsonObjects.AsArray())
     {
         foreach (var key in jsonObject["configuration_keys"].AsArray())
         {
+            if (!string.IsNullOrWhiteSpace(key["secret"].GetValue<string>()))
+            {
 #>
-        {
-            "key": "FoundationaLLM:<#=jsonObject["namespace"]#>:<#=key["name"]#>",
-            "value": "<#=getValue(key)#>",
-            "label": null,
-            "content_type": "<#=getContentType(key)#>",
-            "tags": {}
-        },
+
+        /// <summary>
+        /// <#=key["description"].GetValue<string>()#>
+        /// </summary>
+        public const string FoundationaLLM_<#=getName(jsonObject["namespace"])#>_<#=getName(key["name"])#> =
+            "foundationallm-<#=key["secret"].GetValue<string>()#>";
 <#
+            }
         }
     }
 #>
-    ]
+    }
 }
