+++ src/dotnet/Common/Constants/Configuration/AppConfigurationKeySections.tt
@@ -1,0 +1,40 @@
+<#@ template debug="false" hostspecific="true" language="C#" #>
+<#@ assembly name="System.Core" #>
+<#@ import namespace="System.IO" #>
+<#@ assembly name="System.Text.Json" #>
+<#@ import namespace="System.Collections.Generic" #>
+<#@ assembly name="System.Memory" #>
+<#@ import namespace="System.Text.Json" #>
+<#@ import namespace="System.Text.Json.Nodes" #>
+<#@ output extension=".cs" #>
+<#
+    string inputFileName = this.Host.ResolvePath($"../Data/AppConfiguration.json");
+    string inputContent = File.ReadAllText(inputFileName);
+    var jsonObjects = JsonNode.Parse(inputContent)!;
+    Func<JsonNode, string> getName = n => n.GetValue<string>().Replace(":", "_");
+#>
+namespace FoundationaLLM.Common.Constants.Configuration
+{
+    /// <summary>
+    /// Defines all configuration section names used to map configuration values to settings classes.
+    /// </summary>
+    public static partial class AppConfigurationKeySections
+    {<#
+    foreach (var jsonObject in jsonObjects.AsArray())
+    {
+        var configurationSection = jsonObject["configuration_section"];
+        if (configurationSection != null)
+        {
+#>
+        
+        /// <summary>
+        /// <#=configurationSection["description"].GetValue<string>()#>
+        /// </summary>
+        public const string FoundationaLLM_<#=getName(jsonObject["namespace"])#> =
+            "FoundationaLLM:<#=jsonObject["namespace"]#>";
+<#
+        }
+    }
+#>
+    }
+}
