+++ src/dotnet/Common/Constants/Configuration/AppConfigurationKeys.tt
@@ -1,0 +1,48 @@
+<#@ template debug="false" hostspecific="true" language="C#" #>
+<#@ assembly name="System.Core" #>
+<#@ import namespace="System.IO" #>
+<#@ assembly name="System.Text.Json" #>
+<#@ import namespace="System.Collections.Generic" #>
+<#@ assembly name="System.Memory" #>
+<#@ import namespace="System.Text.Json" #>
+<#@ import namespace="System.Text.Json.Nodes" #>
+<#@ output extension=".cs" #>
+<#
+    string inputFileName = this.Host.ResolvePath($"../Data/AppConfiguration.json");
+    string inputContent = File.ReadAllText(inputFileName);
+    var jsonObjects = JsonNode.Parse(inputContent)!;
+    Func<JsonNode, string> getName = n => n.GetValue<string>().Replace(":", "_");
+#>
+namespace FoundationaLLM.Common.Constants.Configuration
+{
+    /// <summary>
+    /// Defines all App Configuration key names used by FoundationaLLM.
+    /// </summary>
+    public static class AppConfigurationKeys
+    {<#
+    foreach (var jsonObject in jsonObjects.AsArray())
+    {
+#>
+
+        #region FoundationaLLM:<#=jsonObject["namespace"]#>
+<#
+        foreach (var key in jsonObject["configuration_keys"].AsArray())
+        {
+#>
+        
+        /// <summary>
+        /// The app configuration key for the FoundationaLLM:<#=jsonObject["namespace"]#>:<#=key["name"]#> setting.
+        /// <para>Value description:<br/><#=key["description"]#></para>
+        /// </summary>
+        public const string FoundationaLLM_<#=getName(jsonObject["namespace"])#>_<#=getName(key["name"])#> =
+            "FoundationaLLM:<#=jsonObject["namespace"]#>:<#=key["name"]#>";
+<#
+        }
+#>
+
+        #endregion
+<#
+    }
+#>
+    }
+}
