+++ src/dotnet/Common/Constants/Configuration/DependencyInjectionKeys.tt
@@ -1,0 +1,40 @@
+<#@ template debug="false" hostspecific="true" language="C#" #>
+<#@ assembly name="System.Core" #>
+<#@ import namespace="System.IO" #>
+<#@ assembly name="System.Text.Json" #>
+<#@ import namespace="System.Collections.Generic" #>
+<#@ assembly name="System.Memory" #>
+<#@ import namespace="System.Text.Json" #>
+<#@ import namespace="System.Text.Json.Nodes" #>
+<#@ output extension=".cs" #>
+<#
+    string inputFileName = this.Host.ResolvePath($"../Data/AppConfiguration.json");
+    string inputContent = File.ReadAllText(inputFileName);
+    var jsonObjects = JsonNode.Parse(inputContent)!;
+    Func<JsonNode, string> getName = n => n.GetValue<string>().Replace(":", "_");
+#>
+namespace FoundationaLLM.Common.Constants.Configuration
+{
+    /// <summary>
+    /// Defines all keys used for named dependency injection.
+    /// </summary>
+    public static partial class DependencyInjectionKeys
+    {<#
+    foreach (var jsonObject in jsonObjects.AsArray())
+    {
+        var dependencyInjectionKey = jsonObject["dependency_injection_key"];
+        if (dependencyInjectionKey != null)
+        {
+#>
+        
+        /// <summary>
+        /// <#=dependencyInjectionKey["description"].GetValue<string>()#>
+        /// </summary>
+        public const string FoundationaLLM_<#=getName(jsonObject["namespace"])#> =
+            "FoundationaLLM:<#=jsonObject["namespace"]#>";
+<#
+        }
+    }
+#>
+    }
+}
