+++ src/dotnet/Common/Constants/Configuration/KeyVaultSecretNames.cs
@@ -1,183 +1,188 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
 namespace FoundationaLLM.Common.Constants.Configuration
 {
     /// <summary>
-    /// Contains constants for the names of all Azure Key Vault secrets.
+    /// Defines all Azure Key vault secret names referred by the Azure App Configuration keys.
     /// </summary>
-    public static class KeyVaultSecretNames
+    public static partial class KeyVaultSecretNames
     {
         /// <summary>
-        /// The foundationallm-apis-orchestrationapi-apikey Key Vault secret name.
+        /// The name of the Azure Key Vault secret holding the connection string for the App Insights service used by the Core API.
         /// </summary>
-        public const string FoundationaLLM_APIs_OrchestrationAPI_APIKey = "foundationallm-apis-orchestrationapi-apikey";
-        /// <summary>
-        /// The foundationallm-app-insights-connection-string Key Vault secret name.
-        /// </summary>
-        public const string FoundationaLLM_App_Insights_Connection_String = "foundationallm-app-insights-connection-string";
-        /// <summary>
-        /// The foundationallm-apis-agenthubapi-apikey Key Vault secret name.
-        /// </summary>
-        public const string FoundationaLLM_APIs_AgentHubAPI_APIKey = "foundationallm-apis-agenthubapi-apikey";
-        /// <summary>
-        /// The foundationallm-apis-datasourcehubapi-apikey Key Vault secret name.
-        /// </summary>
-        public const string FoundationaLLM_APIs_DataSourceHubAPI_APIKey = "foundationallm-apis-datasourcehubapi-apikey";
-        /// <summary>
-        /// The foundationallm-apis-gatekeeperapi-apikey Key Vault secret name.
-        /// </summary>
-        public const string FoundationaLLM_APIs_GatekeeperAPI_APIKey = "foundationallm-apis-gatekeeperapi-apikey";
+        public const string FoundationaLLM_APIEndpoints_CoreAPI_AppInsightsConnectionString =
+            "foundationallm-appinsights-connection-string";
 
         /// <summary>
-        /// The foundationallm-apis-gatekeeper-lakeraguard-apikey Key Vault secret name.
+        /// The name of the Azure Key Vault secret holding the client secret associated with the Entra ID app registration used by the Core API to authenticate.
         /// </summary>
-        public const string FoundationaLLM_APIs_Gatekeeper_LakeraGuard_APIKey = "foundationallm-apis-gatekeeper-lakeraguard-apikey";
-        /// <summary>
-        /// The foundationallm-apis-gatekeeper-enkryptguardrails-apikey Key Vault secret name.
-        /// </summary>
-        public const string FoundationaLLM_APIs_Gatekeeper_EnkryptGuardrails_APIKey = "foundationallm-apis-gatekeeper-enkryptguardrails-apikey";
-        /// <summary>
-        /// The foundationallm-apis-gatekeeper-azurecontentsafety-apikey Key Vault secret name.
-        /// </summary>
-        public const string FoundationaLLM_APIs_Gatekeeper_AzureContentSafety_APIKey = "foundationallm-apis-gatekeeper-azurecontentsafety-apikey";
+        public const string FoundationaLLM_APIEndpoints_CoreAPI_Configuration_Entra_ClientSecret =
+            "foundationallm-apiendpoints-coreapi-configuration-entra-clientsecret";
 
         /// <summary>
-        /// The foundationallm-apis-gatekeeperintegrationapi-apikey Key Vault secret name.
+        /// The name of the Azure Key Vault secret holding the API key for the Core Worker service.
         /// </summary>
-        public const string FoundationaLLM_APIs_GatekeeperIntegrationAPI_APIKey = "foundationallm-apis-gatekeeperintegrationapi-apikey";
-        /// <summary>
-        /// The foundationallm-apis-langchainapi-apikey Key Vault secret name.
-        /// </summary>
-        public const string FoundationaLLM_APIs_LangChainAPI_APIKey = "foundationallm-apis-langchainapi-apikey";
-        /// <summary>
-        /// The foundationallm-apis-prompthubapi-apikey Key Vault secret name.
-        /// </summary>
-        public const string FoundationaLLM_APIs_PromptHubAPI_APIKey = "foundationallm-apis-prompthubapi-apikey";
-        /// <summary>
-        /// The foundationallm-apis-semantickernelapi-apikey Key Vault secret name.
-        /// </summary>
-        public const string FoundationaLLM_APIs_SemanticKernelAPI_APIKey = "foundationallm-apis-semantickernelapi-apikey";
-        /// <summary>
-        /// The foundationallm-azureopenai-api-key Key Vault secret name.
-        /// </summary>
-        public const string FoundationaLLM_AzureOpenAI_Api_Key = "foundationallm-azureopenai-api-key";
-        /// <summary>
-        /// The foundationallm-chat-entra-clientsecret Key Vault secret name.
-        /// </summary>
-        public const string FoundationaLLM_Chat_Entra_ClientSecret = "foundationallm-chat-entra-clientsecret";
-        /// <summary>
-        /// The foundationallm-coreapi-entra-clientsecret Key Vault secret name.
-        /// </summary>
-        public const string FoundationaLLM_Coreapi_Entra_ClientSecret = "foundationallm-coreapi-entra-clientsecret";
-        /// <summary>
-        /// The foundationallm-langchain-csvfile-url Key Vault secret name.
-        /// </summary>
-        public const string FoundationaLLM_LangChain_CsvFile_Url = "foundationallm-langchain-csvfile-url";
-        /// <summary>
-        /// The foundationallm-langchain-sqldatabase-testdb-password Key Vault secret name.
-        /// </summary>
-        public const string FoundationaLLM_LangChain_SQLDatabase_Testdb_Password = "foundationallm-langchain-sqldatabase-testdb-password";
-        /// <summary>
-        /// The foundationallm-management-entra-clientsecret Key Vault secret name.
-        /// </summary>
-        public const string FoundationaLLM_Management_Entra_ClientSecret = "foundationallm-management-entra-clientsecret";
-        /// <summary>
-        /// The foundationallm-managementapi-entra-clientsecret Key Vault secret name.
-        /// </summary>
-        public const string FoundationaLLM_Managementapi_Entra_ClientSecret = "foundationallm-managementapi-entra-clientsecret";
-        /// <summary>
-        /// The foundationallm-openai-api-key Key Vault secret name.
-        /// </summary>
-        public const string FoundationaLLM_OpenAI_Api_Key = "foundationallm-openai-api-key";
-        /// <summary>
-        /// The foundationallm-semantickernelapi-openai-key Key Vault secret name.
-        /// </summary>
-        public const string FoundationaLLM_SemanticKernelAPI_OpenAI_Key = "foundationallm-semantickernelapi-openai-key";
-        /// <summary>
-        /// The foundationallm-apis-vectorizationapi-apikey Key Vault secret name.
-        /// The API key of the vectorization API.
-        /// </summary>
-        public const string FoundationaLLM_APIs_VectorizationAPI_APIKey = "foundationallm-apis-vectorizationapi-apikey";
-        /// <summary>
-        /// The foundationallm-apis-vectorizationworker-apikey Key Vault secret name.
-        /// The API key of the vectorization worker API.
-        /// </summary>
-        public const string FoundationaLLM_APIs_VectorizationWorker_APIKey = "foundationallm-apis-vectorizationworker-apikey";
-        /// <summary>
-        /// The foundationallm-events-azureeventgrid-apikey Key Vault secret name.
-        /// The API key for the Azure Event Grid service.
-        /// </summary>
-        public const string FoundationaLLM_Events_AzureEventGrid_APIKey = "foundationallm-events-azureeventgrid-apikey";
-        /// <summary>
-        /// The foundationallm-apis-gatewayapi-apikey Key Vault secret name.
-        /// The API key of the Gateway API.
-        /// </summary>
-        public const string FoundationaLLM_APIs_GatewayAPI_APIKey = "foundationallm-apis-gatewayapi-apikey";
-        /// <summary>
-        /// The foundationallm-apis-gatewayadapterapi-apikey Key Vault secret name.
-        /// The API key of the Gateway API.
-        /// </summary>
-        public const string FoundationaLLM_APIs_GatewayAdapterAPI_APIKey = "foundationallm-apis-gatewayadapterapi-apikey";
-        /// <summary>
-        /// The foundationallm-apis-stateapi-apikey Key Vault secret name.
-        /// The API key of the Gateway API.
-        /// </summary>
-        public const string FoundationaLLM_APIs_StateAPI_APIKey = "foundationallm-apis-stateapi-apikey";
-
-        #region Authorization API
+        public const string FoundationaLLM_APIEndpoints_CoreWorker_APIKey =
+            "foundationallm-apiendpoints-coreworker-apikey";
 
         /// <summary>
-        /// The foundationallm-authorizationapi-appinsights-connectionstring Key Vault secret name.
-        /// The connection string used by OpenTelemetry to connect to App Insights.
+        /// The name of the Azure Key Vault secret holding the connection string for the App Insights service used by the Core Worker service.
         /// </summary>
-        public const string FoundationaLLM_AuthorizationAPI_AppInsights_ConnectionString = "foundationallm-authorizationapi-appinsights-connectionstring";
+        public const string FoundationaLLM_APIEndpoints_CoreWorker_AppInsightsConnectionString =
+            "foundationallm-appinsights-connection-string";
 
         /// <summary>
-        /// The foundationallm-authorizationapi-entra- Key Vault secret name.
-        /// The Entra ID instance.
+        /// The name of the Azure Key Vault secret holding the API key for the Gatekeeper API.
         /// </summary>
-        public const string FoundationaLLM_AuthorizationAPI_Entra_Instance = "foundationallm-authorizationapi-entra-instance";
+        public const string FoundationaLLM_APIEndpoints_GatekeeperAPI_APIKey =
+            "foundationallm-apiendpoints-gatekeeperapi-apikey";
 
         /// <summary>
-        /// The foundationallm-authorizationapi-entra- Key Vault secret name.
-        /// The Entra ID tenant id.
+        /// The name of the Azure Key Vault secret holding the API key for the Gatekeeper Integration API.
         /// </summary>
-        public const string FoundationaLLM_AuthorizationAPI_Entra_TenantId = "foundationallm-authorizationapi-entra-tenantid";
+        public const string FoundationaLLM_APIEndpoints_GatekeeperIntegrationAPI_APIKey =
+            "foundationallm-apiendpoints-gatekeeperintergrationapi-apikey";
 
         /// <summary>
-        /// The foundationallm-authorizationapi-entra- Key Vault secret name.
-        /// Entra ID client id.
+        /// The name of the Azure Key Vault secret holding the connection string for the App Insights service used by the Gatekeeper Integration API.
         /// </summary>
-        public const string FoundationaLLM_AuthorizationAPI_Entra_ClientId = "foundationallm-authorizationapi-entra-clientid";
+        public const string FoundationaLLM_APIEndpoints_GatekeeperIntegrationAPI_AppInsightsConnectionString =
+            "foundationallm-appinsights-connection-string";
 
         /// <summary>
-        /// The foundationallm-authorizationapi-entra- Key Vault secret name.
-        /// The Entr ID client secret.
+        /// The name of the Azure Key Vault secret holding the API key for the Orchestration API.
         /// </summary>
-        public const string FoundationaLLM_AuthorizationAPI_Entra_ClientSecret = "foundationallm-authorizationapi-entra-clientsecret";
+        public const string FoundationaLLM_APIEndpoints_OrchestrationAPI_APIKey =
+            "foundationallm-apiendpoints-orchestrationapi-apikey";
 
         /// <summary>
-        /// The foundationallm-authorizationapi-entra- Key Vault secret name.
-        /// The Entra ID scopes.
+        /// The name of the Azure Key Vault secret holding the connection string for the App Insights service used by the Orchestration API.
         /// </summary>
-        public const string FoundationaLLM_AuthorizationAPI_Entra_Scopes = "foundationallm-authorizationapi-entra-scopes";
+        public const string FoundationaLLM_APIEndpoints_OrchestrationAPI_AppInsightsConnectionString =
+            "foundationallm-appinsights-connection-string";
 
         /// <summary>
-        /// The foundationallm-authorizationapi-entra- Key Vault secret name.
-        /// The name of the storage account used by the Authorization API.
+        /// The name of the Azure Key Vault secret holding the API key for the LangChain API.
         /// </summary>
-        public const string FoundationaLLM_AuthorizationAPI_Storage_AccountName = "foundationallm-authorizationapi-storage-accountname";
+        public const string FoundationaLLM_APIEndpoints_LangChainAPI_APIKey =
+            "foundationallm-apiendpoints-langchainapi-apikey";
 
         /// <summary>
-        /// The foundationallm-authorizationapi-entra- Key Vault secret name.
-        /// The comma separated list of the identifiers of FoundationaLLM instances managed by the authorization core.
+        /// The name of the Azure Key Vault secret holding the connection string for the App Insights service used by the LangChain API.
         /// </summary>
-        public const string FoundationaLLM_AuthorizationAPI_InstanceIds = "foundationallm-authorizationapi-instanceids";
+        public const string FoundationaLLM_APIEndpoints_LangChainAPI_AppInsightsConnectionString =
+            "foundationallm-appinsights-connection-string";
 
-        #endregion
+        /// <summary>
+        /// The name of the Azure Key Vault secret holding the API key for the Semantic Kernel API.
+        /// </summary>
+        public const string FoundationaLLM_APIEndpoints_SemanticKernelAPI_APIKey =
+            "foundationallm-apiendpoints-semantickernelapi-apikey";
+
+        /// <summary>
+        /// The name of the Azure Key Vault secret holding the connection string for the App Insights service used by the Semantic Kernel API.
+        /// </summary>
+        public const string FoundationaLLM_APIEndpoints_SemanticKernelAPI_AppInsightsConnectionString =
+            "foundationallm-appinsights-connection-string";
+
+        /// <summary>
+        /// The name of the Azure Key Vault secret holding the connection string for the App Insights service used by the Management API.
+        /// </summary>
+        public const string FoundationaLLM_APIEndpoints_ManagementAPI_AppInsightsConnectionString =
+            "foundationallm-appinsights-connection-string";
+
+        /// <summary>
+        /// The name of the Azure Key Vault secret holding the client secret associated with the Entra ID app registration used by the Management API to authenticate.
+        /// </summary>
+        public const string FoundationaLLM_APIEndpoints_ManagementAPI_Configuration_Entra_ClientSecret =
+            "foundationallm-apiendpoints-managementapi-configuration-entra-clientsecret";
+
+        /// <summary>
+        /// The name of the Azure Key Vault secret holding the API key for the Vectorization API.
+        /// </summary>
+        public const string FoundationaLLM_APIEndpoints_VectorizationAPI_APIKey =
+            "foundationallm-apiendpoints-vectorizationapi-apikey";
+
+        /// <summary>
+        /// The name of the Azure Key Vault secret holding the connection string for the App Insights service used by the Vectorization API.
+        /// </summary>
+        public const string FoundationaLLM_APIEndpoints_VectorizationAPI_AppInsightsConnectionString =
+            "foundationallm-appinsights-connection-string";
+
+        /// <summary>
+        /// The name of the Azure Key Vault secret holding the API key for the Vectorization Worker service.
+        /// </summary>
+        public const string FoundationaLLM_APIEndpoints_VectorizationWorker_APIKey =
+            "foundationallm-apiendpoints-vectorizationworker-apikey";
+
+        /// <summary>
+        /// The name of the Azure Key Vault secret holding the connection string for the App Insights service used by the Vectorization worker service.
+        /// </summary>
+        public const string FoundationaLLM_APIEndpoints_VectorizationWorker_AppInsightsConnectionString =
+            "foundationallm-appinsights-connection-string";
+
+        /// <summary>
+        /// The name of the Azure Key Vault secret holding the API key for the Gateway API.
+        /// </summary>
+        public const string FoundationaLLM_APIEndpoints_GatewayAPI_APIKey =
+            "foundationallm-apiendpoints-gatewayapi-apikey";
+
+        /// <summary>
+        /// The name of the Azure Key Vault secret holding the connection string for the App Insights service used by the Gateway API.
+        /// </summary>
+        public const string FoundationaLLM_APIEndpoints_GatewayAPI_AppInsightsConnectionString =
+            "foundationallm-appinsights-connection-string";
+
+        /// <summary>
+        /// The name of the Azure Key Vault secret holding the API key for the Gateway Adapter API.
+        /// </summary>
+        public const string FoundationaLLM_APIEndpoints_GatewayAdapterAPI_APIKey =
+            "foundationallm-apiendpoints-gatewayadapterapi-apikey";
+
+        /// <summary>
+        /// The name of the Azure Key Vault secret holding the connection string for the App Insights service used by the Gateway Adapter API.
+        /// </summary>
+        public const string FoundationaLLM_APIEndpoints_GatewayAdapterAPI_AppInsightsConnectionString =
+            "foundationallm-appinsights-connection-string";
+
+        /// <summary>
+        /// The name of the Azure Key Vault secret holding the API key for the State API.
+        /// </summary>
+        public const string FoundationaLLM_APIEndpoints_StateAPI_APIKey =
+            "foundationallm-apinedpoints-stateapi-apikey";
+
+        /// <summary>
+        /// The name of the Azure Key Vault secret holding the connection string for the App Insights service used by the State API.
+        /// </summary>
+        public const string FoundationaLLM_APIEndpoints_StateAPI_AppInsightsConnectionString =
+            "foundationallm-appinsights-connection-string";
+
+        /// <summary>
+        /// The name of the Azure Key Vault secret holding the API key for the Azure OpenAI service.
+        /// </summary>
+        public const string FoundationaLLM_APIEndpoints_AzureOpenAI_APIKey =
+            "foundationallm-apiendpoints-azureopenai-apikey";
+
+        /// <summary>
+        /// The name of the Azure Key Vault secret holding the API key for the Azure Event Grid service.
+        /// </summary>
+        public const string FoundationaLLM_APIEndpoints_AzureEventGrid_APIKey =
+            "foundationallm-apiendpoints-azureeventgrid-apikey";
+
+        /// <summary>
+        /// The name of the Azure Key Vault secret holding the API key for the Azure Content Safety service.
+        /// </summary>
+        public const string FoundationaLLM_APIEndpoints_AzureContentSafety_APIKey =
+            "foundationallm-apiendpoints-azurecontentsafety-apikey";
+
+        /// <summary>
+        /// The name of the Azure Key Vault secret holding the client secret associated with the Entra ID app registration used by the User Portal to authenticate.
+        /// </summary>
+        public const string FoundationaLLM_UserPortal_Authentication_Entra_ClientSecret =
+            "foundationallm-userportal-authentication-entra-clientsecret";
+
+        /// <summary>
+        /// The name of the Azure Key Vault secret holding the client secret associated with the Entra ID app registration used by the Management Portal to authenticate.
+        /// </summary>
+        public const string FoundationaLLM_ManagementPortal_Authentication_Entra_ClientSecret =
+            "foundationallm-managementportal-authentication-entra-clientsecret";
     }
 }
