+++ src/dotnet/Common/Constants/Configuration/KeyVaultSecretNames.tt
@@ -11,37 +11,36 @@
     string inputFileName = this.Host.ResolvePath($"../Data/AppConfiguration.json");
     string inputContent = File.ReadAllText(inputFileName);
     var jsonObjects = JsonNode.Parse(inputContent)!;
+
     Func<JsonNode, string> getName = n => n.GetValue<string>().Replace(":", "_");
+
+    Func<JsonNode, string> getValue = n => string.IsNullOrWhiteSpace(n["secret"].GetValue<string>())
+        ? n["value"].GetValue<string>()
+        : $"{{\\\"uri\\\":\\\"${{env:AZURE_KEY_VAULT_ENDPOINT}}secrets/foundationallm-{n["secret"].GetValue<string>()}}}\\\"}}";
 #>
 namespace FoundationaLLM.Common.Constants.Configuration
 {
     /// <summary>
-    /// Defines all App Configuration key names used by FoundationaLLM.
+    /// Defines all Azure Key vault secret names referred by the Azure App Configuration keys.
     /// </summary>
-    public static class AppConfigurationKeysNew
+    public static partial class KeyVaultSecretNames
     {<#
     foreach (var jsonObject in jsonObjects.AsArray())
     {
+        foreach (var key in jsonObject["configuration_keys"].AsArray())
+        {
+            if (!string.IsNullOrWhiteSpace(key["secret"].GetValue<string>()))
+            {
 #>
 
-        #region FoundationaLLM:<#=jsonObject["namespace"]#>
-<#
-        foreach (var key in jsonObject["keys"].AsArray())
-        {
-#>
-        
         /// <summary>
-        /// The app configuration key for the FoundationaLLM:<#=jsonObject["namespace"]#>:<#=key["name"]#> setting.
-        /// <para>Value description:<br/><#=key["description"]#></para>
+        /// <#=key["description"].GetValue<string>()#>
         /// </summary>
         public const string FoundationaLLM_<#=getName(jsonObject["namespace"])#>_<#=getName(key["name"])#> =
-            "FoundationaLLM:<#=jsonObject["namespace"]#>:<#=key["name"]#>";
+            "foundationallm-<#=key["secret"].GetValue<string>()#>";
 <#
+            }
         }
-#>
-
-        #endregion
-<#
     }
 #>
     }
