+++ src/dotnet/Common/Templates/authorization-azure-key-vault-secret-names.json
@@ -1,0 +1,8 @@
+{
+    "FoundationaLLM:ResourceProviders:Authorization:Storage:AccountName": "foundationallm-authorizationapi-storage-accountname",
+    "FoundationaLLM:APIEndpoints:AuthorizationAPI:AppInsightsConnectionString": "foundationallm-authorizationapi-appinsights-connectionstring",
+    "FoundationaLLM:APIEndpoints:AuthorizationAPI:Configuration:InstanceIds": "foundationallm-authorizationapi-instanceids",
+    "FoundationaLLM:APIEndpoints:AuthorizationAPI:Configuration:Entra:Instance": "foundationallm-authorizationapi-entra-instance",
+    "FoundationaLLM:APIEndpoints:AuthorizationAPI:Configuration:Entra:TenantId": "foundationallm-authorizationapi-entra-tenantid",
+    "FoundationaLLM:APIEndpoints:AuthorizationAPI:Configuration:Entra:ClientId": "foundationallm-authorizationapi-entra-clientid",
+}
