+++ src/dotnet/Common/Templates/authorization-azure-key-vault-secret-names.tt
@@ -8,7 +8,7 @@
 <#@ import namespace="System.Text.Json.Nodes" #>
 <#@ output extension=".json" #>
 <#
-    string inputFileName = this.Host.ResolvePath($"../Constants/Data/AppConfiguration.json");
+    string inputFileName = this.Host.ResolvePath($"../Constants/Data/AuthorizationAppConfiguration.json");
     string inputContent = File.ReadAllText(inputFileName);
     var jsonObjects = JsonNode.Parse(inputContent)!;
 
@@ -21,23 +21,18 @@
         : $"{{\\\"uri\\\":\\\"${{env:AZURE_KEY_VAULT_ENDPOINT}}secrets/foundationallm-{n["secret"].GetValue<string>()}}}\\\"}}";
 #>
 {
-    "items": [
 <#
     foreach (var jsonObject in jsonObjects.AsArray())
     {
         foreach (var key in jsonObject["configuration_keys"].AsArray())
         {
+            if (!string.IsNullOrWhiteSpace(key["secret"].GetValue<string>()))
+            {
 #>
-        {
-            "key": "FoundationaLLM:<#=jsonObject["namespace"]#>:<#=key["name"]#>",
-            "value": "<#=getValue(key)#>",
-            "label": null,
-            "content_type": "<#=getContentType(key)#>",
-            "tags": {}
-        },
+    "FoundationaLLM:<#=jsonObject["namespace"]#>:<#=key["name"]#>": "<#=$"foundationallm-{key["secret"].GetValue<string>()}"#>",
 <#
+            }
         }
     }
 #>
-    ]
 }
