+++ src/dotnet/Common/Templates/azure-key-vault-secret-names.tt
@@ -24,7 +24,7 @@
 <#
     foreach (var jsonObject in jsonObjects.AsArray())
     {
-        foreach (var key in jsonObject["keys"].AsArray())
+        foreach (var key in jsonObject["configuration_keys"].AsArray())
         {
             if (!string.IsNullOrWhiteSpace(key["secret"].GetValue<string>()))
             {
