+++ src/dotnet/Configuration/Services/ConfigurationResourceProviderService.cs
@@ -40,7 +40,7 @@ namespace FoundationaLLM.Configuration.Services
     public class ConfigurationResourceProviderService(
         IOptions<InstanceSettings> instanceOptions,
         IAuthorizationService authorizationService,
-        [FromKeyedServices(DependencyInjectionKeys.FoundationaLLM_ResourceProvider_Configuration)] IStorageService storageService,
+        [FromKeyedServices(DependencyInjectionKeys.FoundationaLLM_ResourceProviders_Configuration)] IStorageService storageService,
         IEventService eventService,
         IResourceValidatorFactory resourceValidatorFactory,
         IAzureAppConfigurationService appConfigurationService,
