+++ tests/dotnet/Core.Examples/Setup/TestServicesInitializer.cs
@@ -43,7 +43,7 @@ namespace FoundationaLLM.Core.Examples.Setup
 			TestConfiguration.Initialize(configRoot, services);
 
             services.AddOptions<BlobStorageServiceSettings>(
-                    DependencyInjectionKeys.FoundationaLLM_ResourceProvider_Vectorization)
+                    DependencyInjectionKeys.FoundationaLLM_ResourceProviders_Vectorization)
                 .Bind(configRoot.GetSection("FoundationaLLM:Vectorization:ResourceProviderService:Storage"));
 
             RegisterInstance(services, configRoot);
